backend1/backendtest/requirements.txt
dgl
dgllife
Flask
matplotlib
numpy
pandas
Pillow
scikit_learn
scipy
tqdm
joblib
flask_cors

pip3 install Flask matplotlib numpy pandas Pillow scikit_learn scipy tqdm joblib flask_cors



# conda install -c rdkit rdkit==2018.09.3
# pip3 install torch torchvision torchaudio
# pip3 install dgl
# pip3 install dgllife